PYTHON/requirements.txt
Flask==3.0.2
PyPDF2==3.0.1
python-dotenv==1.0.1
spacy==3.7.4
pdfminer.six==20231228
pdf2image==1.17.0
pytesseract==0.3.10
pillow==10.2.0
regex==2023.12.25
redis==5.0.1
python-dotenv==1.0.1
flask-paginate==2024.4.12
python-dateutil==2.8.0
datefinder==0.7.3
python-docx==1.2.0
PyJWT==2.10.1
bcrypt==4.3.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl